Code cleanup, guys. Because hackathon.

+++ Mesh/WSMMessagingViewController.m
@@ -6,10 +6,11 @@
 //  Copyright (c) 2014 wrkstrm. All rights reserved.
 //
 
+#import <SOMessaging/SOMessage.h>
+#import <LayerKit/LayerKit.h>
 #import "WSMMessagingViewController.h"
 #import "ContentManager.h"
 #import "Message.h"
-
 
 @interface WSMMessagingViewController ()
 
@@ -21,19 +22,14 @@
 
 @implementation WSMMessagingViewController
 
-
-- (void)viewDidLoad
-{
+- (void)viewDidLoad {
     [super viewDidLoad];
-    // Do any additional setup after loading the view.
-
     NSData *meContent, *recipientContent;
     meContent = [[self.currentUser attachmentNamed:@"avatar"] content];
     recipientContent = [[self.recipient attachmentNamed:@"avatar"] content];
     
-    self.myImage      = [UIImage imageWithData:meContent];
+    self.myImage = [UIImage imageWithData:meContent];
     self.partnerImage = [UIImage imageWithData:recipientContent];
-    
     
     [self loadMessages];
 }
@@ -42,26 +38,21 @@
     return [[WSMUserManager sharedInstance] currentUser];
 }
 
-- (void)loadMessages
-{
+- (void)loadMessages {
     self.dataSource = [[[ContentManager sharedManager] generateConversation] mutableCopy];
 }
 
 #pragma mark - SOMessaging data source
-- (NSMutableArray *)messages
-{
+- (NSMutableArray *)messages {
     return self.dataSource;
 }
 
-- (NSTimeInterval)intervalForMessagesGrouping
-{
-    // Return 0 for disableing grouping
-    return 2 * 24 * 3600;
+- (NSTimeInterval)intervalForMessagesGrouping {
+    return 2 * 24 * 3600; // Return 0 for disableing grouping
 }
 
-- (void)configureMessageCell:(SOMessageCell *)cell forMessageAtIndex:(NSInteger)index
-{
-    Message *message = self.dataSource[index];
+- (void)configureMessageCell:(SOMessageCell *)cell forMessageAtIndex:(NSInteger)index {
+    Message *message = self.dataSource[(NSUInteger)index];
     
     // Adjusting content for 3pt. (In this demo the width of bubble's tail is 3pt)
     if (!message.fromMe) {
@@ -81,39 +72,31 @@
     cell.userImage = message.fromMe ? self.myImage : self.partnerImage;
 }
 
-- (CGFloat)messageMaxWidth
-{
+- (CGFloat)messageMaxWidth {
     return 140;
 }
 
-- (CGSize)userImageSize
-{
+- (CGSize)userImageSize {
     return CGSizeMake(40, 40);
 }
 
-- (CGFloat)messageMinHeight
-{
+- (CGFloat)messageMinHeight {
     return 0;
 }
+
 #pragma mark - SOMessaging delegate
-- (void)didSelectMedia:(NSData *)media inMessageCell:(SOMessageCell *)cell
-{
-    // Show selected media in fullscreen
+- (void)didSelectMedia:(NSData *)media inMessageCell:(SOMessageCell *)cell {
     [super didSelectMedia:media inMessageCell:cell];
 }
 
-- (void)messageInputView:(SOMessageInputView *)inputView didSendMessage:(NSString *)message
-{
+- (void)messageInputView:(SOMessageInputView *)inputView didSendMessage:(NSString *)message {
     if (![[message stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]] length]) {
         return;
     }
     
     LYRClient *client = [LYRClient sharedClient];
-    
     LYRAddress *address = [LYRAddress addressWithString:self.recipient.document[@"address"]];
-    
     LYRMessage *outgoingMessage = [[LYRMessage alloc] initWithClient:client];
-    
     
     NSError *error;
     
@@ -133,8 +116,7 @@
     }];
 }
 
-- (void)messageInputViewDidSelectMediaButton:(SOMessageInputView *)inputView
-{
+- (void)messageInputViewDidSelectMediaButton:(SOMessageInputView *)inputView {
     // Take a photo/video or choose from gallery
 }
